Ability to set global fetch policy & implement ssrForceFetchDelay
1. fetch policy is first read from current useData, if it is present,
we use it. If it is not present, we check for policy in global client
if that is also not present, we got for default cache-first

2. ssrForceFetchDelay timer starts when data client is created, during
render we again take new time, see the diff and decide to use from
cache or to be make a netowork.

3. Support for it in HOC yet to be addded, example yet to be updated
and tests needs to be updated - (pending task)